prepare separate `db` folder with db client and definitions
