+++ src/db/user-definitions.ts
@@ -1,0 +1,1 @@
+export const defs = 'user-definitions';
