+++ tsconfig.json
@@ -20,7 +20,8 @@
         ],
         "paths": {
             "@/*": ["./src/*"]
-        }
+        },
+        "baseUrl": "."
     },
     "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
     "exclude": ["node_modules"]
